+++ extension/js/common/browser/browser-msg.ts
@@ -120,7 +120,6 @@ export class BrowserMsg {
 
   public static send = { // todo - may want to organise this differently, seems to always confuse me when sending a message
     bg: {
-      settings: (bm: Bm.Settings) => BrowserMsg.sendCatch(undefined, 'settings', bm),
       updateUninstallUrl: () => BrowserMsg.sendCatch(undefined, 'update_uninstall_url', {}),
       await: {
         reconnectAcctAuthPopup: (bm: Bm.ReconnectAcctAuthPopup) => BrowserMsg.sendAwait(undefined, 'reconnect_acct_auth_popup', bm, true) as Promise<Bm.Res.ReconnectAcctAuthPopup>,
