+++ extension/js/content_scripts/webmail/gmail-element-replacer.ts
@@ -281,8 +281,8 @@ export class GmailElementReplacer implements WebmailElementReplacer {
               }
             } else {
               const statusMsg = 'Missing Gmail permission to decrypt attachments. <a href="#" class="auth_settings">Settings</a></div>';
-              $(newPgpAtts).prepend(this.factory.embeddedAttaStatus(statusMsg)).children('a.auth_settings').click(Ui.event.handle(() => { // xss-safe-factory
-                BrowserMsg.send.bg.settings({ acctEmail: this.acctEmail, page: '/chrome/settings/modules/auth_denied.htm' });
+              $(newPgpAtts).prepend(this.factory.embeddedAttaStatus(statusMsg)).children('a.auth_settings').click(Ui.event.handle(async () => { // xss-safe-factory
+                await Browser.openSettingsPageWithChromeTabs(undefined, this.acctEmail, '/chrome/settings/modules/auth_denied.htm');
               }));
             }
           } else {
@@ -643,7 +643,7 @@ export class GmailElementReplacer implements WebmailElementReplacer {
       const settingsBtnContainer = $(this.sel.settingsBtnContainer);
       if (settingsBtnContainer.length && !settingsBtnContainer.find('#fc_settings_btn').length) {
         settingsBtnContainer.children().last().before(this.factory.btnSettings('gmail')); // xss-safe-factory
-        settingsBtnContainer.find('#fc_settings_btn').click(Ui.event.handle(() => BrowserMsg.send.bg.settings({ acctEmail: this.acctEmail })));
+        settingsBtnContainer.find('#fc_settings_btn').click(Ui.event.handle(async () => await Browser.openSettingsPageWithChromeTabs(undefined, this.acctEmail)));
       }
     }
   }
